Move camp logic to managers.
