+++ src/engine/core/database/registry.ts
@@ -4,6 +4,7 @@ import type {
   IStoredOfflineObject,
 } from "@/engine/core/database/database_types";
 import type { AbstractManager, TAbstractCoreManagerConstructor } from "@/engine/core/managers/base/AbstractManager";
+import type { CampManager } from "@/engine/core/managers/camp";
 import type { TSimulationObject } from "@/engine/core/managers/simulation";
 import type { ITradeManagerDescriptor } from "@/engine/core/managers/trade";
 import type { TAbstractSchemeConstructor } from "@/engine/core/objects/ai/scheme";
@@ -13,7 +14,6 @@ import type {
   LabX8DoorBinder,
   SignalLightBinder,
 } from "@/engine/core/objects/binders";
-import type { CampManager } from "@/engine/core/objects/camp";
 import type { Actor } from "@/engine/core/objects/server/creature/Actor";
 import type { SmartCover } from "@/engine/core/objects/server/smart_cover";
 import type { SmartTerrain } from "@/engine/core/objects/server/smart_terrain";
