+++ src/engine/core/managers/camp/CampManager.test.ts
