+++ src/engine/core/objects/binders/zones/CampZoneBinder.ts
@@ -10,7 +10,7 @@ import {
   resetCampZone,
   unregisterCampZone,
 } from "@/engine/core/database";
-import { CampManager } from "@/engine/core/objects/camp/CampManager";
+import { CampManager } from "@/engine/core/managers/camp/CampManager";
 import { readIniString } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { IniFile, NetPacket, Optional, Reader, ServerObject, TDuration, TName } from "@/engine/lib/types";
