--- src/engine/core/objects/camp/index.ts
@@ -1,3 +1,0 @@
-export * from "@/engine/core/objects/camp/camp_types";
-export * from "@/engine/core/objects/camp/camp_logic";
-export * from "@/engine/core/objects/camp/CampManager";
